requirements.txt
django
psycopg2-binary
djangorestframework
cloudinary
django-cloudinary-storage
gunicorn
